requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
langchain==0.2.10
langchain_community==0.0.14
flask
pypdf
python-dotenv
-e .


